// gen/ts/src/device/descriptor.ts
export interface DeviceDescriptor {
  deviceId?:
    | DeviceId
    | undefined;
  /** Matter 1.5.1 device type ID (e.g. 0x0301 = Thermostat, 0x0302 = TemperatureSensor) */
  matterDeviceTypeId?:
    | number
    | undefined;
  /** Matter BasicInformation cluster attributes */
  vendorName?: string | undefined;
  productName?: string | undefined;
  serialNumber?: string | undefined;
  nodeLabel?: string | undefined;
  vendorId?: number | undefined;
  productId?: number | undefined;
  hardwareVersionString?: string | undefined;
  softwareVersionString?: string | undefined;
  connectorId?: ConnectorId | undefined;
  spaceId?: SpaceId | undefined;
  ownership?: DeviceOwnershipType | undefined;
  lifecycle?:
    | DeviceLifecycleState
    | undefined;
  /** Populated when ownership == RESIDENT */
  ownerUserId?: UserId | undefined;
  claimedAt?: Date | undefined;
}