// gen/ts/src/common/types.ts
export enum DeviceOwnershipType {
  DEVICE_OWNERSHIP_TYPE_UNSPECIFIED = "DEVICE_OWNERSHIP_TYPE_UNSPECIFIED",
  COMPANY = "COMPANY",
  RESIDENT = "RESIDENT",
  UNRECOGNIZED = "UNRECOGNIZED",
}
// ...
export enum DeviceLifecycleState {
  DEVICE_LIFECYCLE_STATE_UNSPECIFIED = "DEVICE_LIFECYCLE_STATE_UNSPECIFIED",
  PENDING_CLAIM = "PENDING_CLAIM",
  OWNED = "OWNED",
  PENDING_TRANSFER = "PENDING_TRANSFER",
  BUILDING_POOL = "BUILDING_POOL",
  UNRECOGNIZED = "UNRECOGNIZED",
}