// gen/ts/src/device/property_update.ts
export interface PropertyUpdate {
  deviceId?: DeviceId | undefined;
  attributeName?: string | undefined;
  value?:
    | { $case: "intValue"; intValue: number }
    | { $case: "uintValue"; uintValue: number }
    | { $case: "boolValue"; boolValue: boolean }
    | { $case: "floatValue"; floatValue: number }
    | { $case: "stringValue"; stringValue: string }
    | { $case: "bytesValue"; bytesValue: Buffer }
    | undefined;
  timestamp?: Date | undefined;
  clusterIdHex?: string | undefined;
}

export interface PropertyUpdateBatch {
  updates?: PropertyUpdate[] | undefined;
  batchTimestamp?: Date | undefined;
}
// ...
export const PropertyUpdateBatch: MessageFns<PropertyUpdateBatch> = {
  encode(message: PropertyUpdateBatch, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.updates !== undefined && message.updates.length !== 0) {
      for (const v of message.updates) {
        PropertyUpdate.encode(v!, writer.uint32(10).fork()).join();
      }
    }
    if (message.batchTimestamp !== undefined) {
      Timestamp.encode(toTimestamp(message.batchTimestamp), writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): PropertyUpdateBatch {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePropertyUpdateBatch();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          const el = PropertyUpdate.decode(reader, reader.uint32());
          if (el !== undefined) {
            message.updates!.push(el);
          }
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.batchTimestamp = fromTimestamp(Timestamp.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PropertyUpdateBatch {
    return {
      updates: globalThis.Array.isArray(object?.updates)
        ? object.updates.map((e: any) => PropertyUpdate.fromJSON(e))
        : [],
      batchTimestamp: isSet(object.batchTimestamp)
        ? fromJsonTimestamp(object.batchTimestamp)
        : isSet(object.batch_timestamp)
        ? fromJsonTimestamp(object.batch_timestamp)
        : undefined,
    };
  },

  toJSON(message: PropertyUpdateBatch): unknown {
    const obj: any = {};
    if (message.updates?.length) {
      obj.updates = message.updates.map((e) => PropertyUpdate.toJSON(e));
    }
    if (message.batchTimestamp !== undefined) {
      obj.batchTimestamp = message.batchTimestamp.toISOString();
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<PropertyUpdateBatch>, I>>(base?: I): PropertyUpdateBatch {
    return PropertyUpdateBatch.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<PropertyUpdateBatch>, I>>(object: I): PropertyUpdateBatch {
    const message = createBasePropertyUpdateBatch();
    message.updates = object.updates?.map((e) => PropertyUpdate.fromPartial(e)) || [];
    message.batchTimestamp = object.batchTimestamp ?? undefined;
    return message;
  },
};

function bytesFromBase64(b64: string): Uint8Array {
  return Uint8Array.from(globalThis.Buffer.from(b64, "base64"));
}

function base64FromBytes(arr: Uint8Array): string {
  return globalThis.Buffer.from(arr).toString("base64");
}

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string } ? { [K in keyof Omit<T, "$case">]?: DeepPartial<T[K]> } & { $case: T["$case"] }
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}